mcp-server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "threat-designer-mcp-server"
version = "1.0.0"
description = "MCP server for Threat Designer API"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
license = {text = "APACHE"}
requires-python = ">=3.10"
dependencies = [
    "boto3",
    "mcp>=1.0.0",
    "fastmcp",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "pillow>=9.0.0",
]

[project.scripts]
threat-designer-mcp-server = "threat_designer_mcp.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["threat_designer_mcp*"]

[project.urls]
Homepage = "https://github.com/awslabs/threat-designer"
Repository = "https://github.com/awslabs/threat-designer"